Move certificate session creation route to dedicated blueprint
