+++ CONTEXT.md
@@ -630,4 +630,5 @@ Route inventory lives at `sitemap.txt` (admin-only, linked from Settings) and li
 - Materials order Special instructions textarea uses the same wide alignment as the session form so its left edge matches other inputs.
 - Materials order header cards (Order details + Shipping details) share a `.materials-header-grid` two-column layout that stacks below 1100px, and the shipping location dropdown lists each location's title only.
 - Certificate-only sessions (`delivery_type = "Certificate only"`) also persist `Session.is_certificate_only = True`, automatically flip Ready for delivery on create/edit, force `no_material_order`, `no_prework`, and `prework_disabled`, hide Materials and Prework navigation (including dashboards and lists), and continue to follow the existing attendance + certificate rules.
-- KT Staff and App Admin can create certificate-only workshops via `GET/POST /certificates/new` (endpoint `certificates.new_certificate_session`). The form captures client, data region, workshop type, language, location, facilitators, schedule, and number of class days, marks the session ready for delivery, and redirects to the session detail page. The left nav surfaces a “New Certificate Session” link directly under “New Order” and before “Workshop Dashboard” for these users, and the page highlights that nav entry when active.
+- KT Staff and App Admin can create certificate-only workshops via `GET/POST /certificate-sessions/new` (endpoint `certificate_sessions.new`). The form captures client, data region, workshop type, language, location, facilitators, schedule, and number of class days, marks the session ready for delivery, and redirects to the session detail page. The left nav surfaces a “New Certificate Session” link directly under “New Order” and before “Workshop Dashboard” for these users, and the page highlights that nav entry when active.
+- `/certificates/*` paths are reserved for static certificate PDFs served by Caddy; the certificate creation UI now lives under `/certificate-sessions/*`.
