+++ tests/smoke/test_auth_roles.py
@@ -78,10 +78,10 @@ def test_auth_roles_home_selection(app, client):
     assert "Switch views here." in admin_html
     assert "<label>View:</label>" in admin_html
     assert "Switch to Admin" not in admin_html
-    cert_form = client.get("/certificates/new")
+    cert_form = client.get("/certificate-sessions/new")
     assert cert_form.status_code == 200
     assert "New Certificate Session" in cert_form.get_data(as_text=True)
-    cert_form_slash = client.get("/certificates/new/")
+    cert_form_slash = client.get("/certificate-sessions/new/")
     assert cert_form_slash.status_code == 200
     response = client.get(
         "/settings/view", query_string={"view": "MATERIAL_MANAGER"}, follow_redirects=False
